requirements.txt
pandas
numpy
lightgbm
scikit-learn
fastapi
uvicorn
python-dotenv
requests
requests
